Refactor : Modal의 폴더 명을 ChannelModal으로 변경
