+++ frontend/src/components/Main/MainPage.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import ServerListTab from './ServerListTab';
 import MainSection from './MainSection';
 import { MainStoreContext } from './MainStore';
-import CreateChannelModal from './Modal/CreateChannelModal';
-import JoinChannelModal from './Modal/JoinChannelModal';
+import CreateChannelModal from './ChannelModal/CreateChannelModal';
+import JoinChannelModal from './ChannelModal/JoinChannelModal';
 import CreateServerModal from './ServerModal/CreateServerModal';
 import JoinServerModal from './ServerModal/JoinServerModal';
 import ServerSettingModal from './ServerModal/ServerSettingModal';
